Improvise a Jazz Solo with an LSTM Network
In this project, I delved into the exciting realm of music generation using deep learning techniques. Leveraging the power of TensorFlow, I implemented a Long Short-Term Memory (LSTM) network to compose jazz solos. The project involved using a variety of tools and libraries, including:

- **TensorFlow:** The core deep learning framework that provided the backbone for implementing the LSTM network for music generation.

- **music21:** A Python toolkit for computer-aided musicology, used for handling musical notation and analyzing musical structures.

- **matplotlib:** Employed for creating visualizations and plots to better understand the model's performance and generated music.

- **NumPy:** Essential for numerical operations and handling arrays, especially in the preprocessing and manipulation of musical data.

- **IPython:** Utilized for interactive computing and testing during the development process.

The neural network architecture was constructed using the Keras API, with layers such as LSTM, Dense, Activation, Dropout, and Reshape contributing to the model's design. The Adam optimizer was employed to train the model, and categorical encoding was utilized for efficient representation of musical data.

The collaboration of these tools facilitated the creation of a powerful system for jazz solo generation, allowing me to explore the intersection of deep learning and musical creativity.